requirements.txt
streamlit
python-dotenv
llama-index  
